Added Background component and changed animations to render globally instead of by page
