+++ StudyWitMe-app/.gitignore
@@ -1,3 +1,5 @@
+.env
+
 # Logs
 logs
 *.log
@@ -22,4 +24,4 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
-StudyWitMe/StudyWitMe-app/src/assets**/logo_files/
+logo_files/
