+++ StudyWitMe-app/src/components/MainMenu.jsx
@@ -9,7 +9,7 @@ function MainMenu() {
 
     return (
         <div className="main-menu">
-        <div className="pattern"></div>
+
 
         <header className="menu-header">
         <nav className="nav-buttons">
@@ -28,7 +28,6 @@ function MainMenu() {
                     </div>
                 )}
             </div>
-
             <button>Play</button>
             <button onClick={() => navigate("/")}>Sign Out</button>
             </nav>
@@ -38,51 +37,6 @@ function MainMenu() {
         <div className="menu-body">
             <h2>Welcome to StudyWitMe!</h2>
             <p>Select an option from the menu above.</p>
-        </div>
-                {/* Background Shapes container */}
-        <div className="shapes">
-            {/* Skinny rectangles */}
-            <div className="shape rectangle bottom-left" />
-            <div className="shape rectangle bottom-right" />
-
-            {/* Hand-placed circles */}
-            <div
-            className="shape circle filled"
-            style={{ width: "40px", height: "40px", left: "10%", animationDelay: "0s" }}
-            />
-            <div
-            className="shape circle hollow"
-            style={{ width: "60px", height: "60px", left: "30%", animationDelay: "2s" }}
-            />
-            <div
-            className="shape circle filled"
-            style={{ width: "30px", height: "30px", left: "70%", animationDelay: "4s" }}
-            />
-            <div
-            className="shape circle hollow"
-            style={{ width: "50px", height: "50px", left: "85%", animationDelay: "6s" }}
-            />
-
-            {/* Randomized circles */}
-            {Array.from({ length: 5 }).map((_, i) => {
-            const size = Math.floor(Math.random() * 40) + 20;
-            const left = Math.random() * 90;
-            const delay = `${(Math.random() * 8).toFixed(2)}s`;
-            const filled = Math.random() > 0.5;
-
-            return (
-                <div
-                key={`circle-${i}`}
-                className={`shape circle ${filled ? "filled" : "hollow"}`}
-                style={{
-                    width: `${size}px`,
-                    height: `${size}px`,
-                    left: `${left}%`,
-                    animationDelay: delay,
-                }}
-                />
-            );
-            })}
         </div>
         </div>
     );
