+++ package.json
@@ -1,7 +1,8 @@
 {
   "dependencies": {
     "@google/generative-ai": "^0.24.1",
-    "express": "^5.1.0"
+    "express": "^5.1.0",
+    "firebase": "^12.3.0"
   },
   "name": "studywitme",
   "version": "1.0.0",
